+++ marketplace/deployer_util/config_helper.py
@@ -1,0 +1,39 @@
+#!/usr/bin/env python2
+#
+# Copyright 2018 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import os
+import re
+
+NAME_RE = re.compile(r'[a-zA-z0-9_]+$')
+
+
+class InvalidName(Exception):
+  pass
+
+
+def read_values_to_dict(values_dir, codec):
+  """Returns a dict constructed from files in values_dir."""
+  files = [f for f in os.listdir(values_dir)
+           if os.path.isfile(os.path.join(values_dir, f))]
+  result = {}
+  for filename in files:
+    if not NAME_RE.match(filename):
+      raise InvalidName('Invalid config parameter name: {}'.format(filename))
+    file_path = os.path.join(values_dir, filename)
+    with open(file_path, "r") as f:
+      data = f.read().decode(codec)
+      result[filename] = data
+  return result
